Fixes bug in voucher generation for items with variations
Bug was for voucher generation to fail when an item with variations was
chosen.
In items with variations, vouchers will be generated if at least one
variation is still available and the sum of available items is larger
than the amount of vouchers being generated.

It would maybe be worth consideration to make this the implementation of
Item.check_quotas for items with variations.

Refs #170.