Event setup: do not create new team when staff session active
With b260cca412df43d6e47d4d574cb2e5132399c2c0, team provisioning has
been disabled for users that are staff. However, this leads to strange
UX, as a new event created by a staff member, not currently in a staff
session resulted in a 404 directly after creation.

The PR updates this requirement to not need to select a team, only when
a staff session is active.