Docs: Make sure data directory is accessible
Make sure that the data directory of small-scale docker deployments is
accessible to the user running nginx inside the container, so static
files can be served from there. This is already the default for most
Linux distributions, because the umask is usually 022, unless it has
been modified inside the current shell.

This change is motivated by #1587.