implement customer_accounts_required setting
Currently, when customers are not logged in, they can still see events,
for example on which date(s) the event is scheduled.

When the customer_accounts_required setting is enabled, customers
always need to login to view any page of this shop.

This is useful for shops with closed user groups and prevents confusion:
We have had a customer not understand how to order, because the
“add to cart” button was missing when the customer is not logged in.
With this setting, we enforce a login, and hence a functional event view.