Corrections to the Order and Examples Defined in PERSON_NAME_SCHEMES
The reasons for the corrections are as follows:

	1.	"family_nospace_given" is originally defined as having the order of “given name, family name,” with the example of “毛泽东” (Mao Zedong) written in simplified Chinese characters.

However, this results in names being displayed as “泽东毛” (= ZedongMao), which is inconsistent with the example and opposite to the name order in cultural contexts where Chinese characters are used.

Furthermore, using the simplified Chinese version of “毛泽东” may be difficult to understand in regions that do not use simplified characters. Additionally, the name of a leader of the Chinese Communist Party may be considered inappropriate in some regions.

I propose replacing this example with “孫文” (Sun Wen, the official name of “Sun Yat-sen”).

The characters in this name are identical in both simplified and traditional Chinese, and they also match the character style used in Japan (Kanji) and Korea (Hanja). Moreover, Sun Yat-sen is regarded as the “Father of the Nation” in both mainland China and Taiwan and is generally recognized as a politically neutral figure.

	2.	"full_transcription" currently lists “庄司” (Shoji) as an example of a full name, but this is clearly incorrect.

“庄司” is the kanji representation of a Japanese surname "Shoji" and is not a full name. According to Japanese law, a full name must include both a family name and a given name. The only exception to this rule applies to members of the Imperial Family.

Therefore, I suggest revising the example to “明仁” (Akihito, the Emperor of Japan).

It should also be noted that in some regions of Asia, legal names consist only of given names without surnames. If pretix plans to support languages from such regions in the future, it would be more appropriate to use examples of full names from those languages in this section. (e.g., Burmese, traditional Mongolian, Javanese, etc.)